+++ 3o14/src/index.ts
@@ -1,6 +1,6 @@
 import { serve } from "@hono/node-server";
 import { behindProxy } from "x-forwarded-fetch";
-import app from "./app.tsx";
+import app from "./routes/index.ts";
 import "./logging.ts";
 
 serve(
